readFile with double just need to especify %lf (long float)

+++ ProgrammingExamples/Cpp/files_managment/readFile.cpp
@@ -6,11 +6,11 @@
 int main(int argc, char* argv[]) {
 	FILE *fp;
 	fp = fopen("data.dat","r");
-	int number;
+	double number;
 	printf("read file\n");
 	for (int i=0;i<6;i++){
-		fscanf(fp,"%d",&number);
-		printf("number = %d\n",number);
+		fscanf(fp,"%lf",&number);
+		printf("number = %f\n",number);
 	}
 	fclose(fp);
 	return EXIT_SUCCESS;
